extract base scan result controller
